Experimental pattern for tcs

+++ common/messages/bin-packets/src/packets/mod.rs
@@ -58,9 +58,9 @@ pub enum ApplicationPacket {
     },
     EnvironmentData {
         timestamp: u64,
-        temperature: u32,
-        pressure: u32,
-        humidity: u16,
+        temperature: f32,
+        pressure: f32,
+        humidity: f32,
     },
     BMPData {
         timestamp: u64,
